fix(base-utils/buffer-and-hex-utils): Fix circular dependencies
